Moved provider classes to another package
